email_assistant/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-cohere>=0.1.0
openai>=1.40.0
cohere>=5.0.0
streamlit>=1.38.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
fpdf2>=2.7.0
pydantic>=2.0.0
typing-extensions>=4.8.0
